refactor(lane_change): separate structs to different folders
